# src/main/java/com/techtest/mobilab/services/api/TransactionApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.19).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.techtest.mobilab.services.api;

import com.techtest.mobilab.services.model.ModelApiResponse;
import com.techtest.mobilab.services.model.Transaction;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;

import java.text.ParseException;
import java.util.List;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2021-04-06T09:18:58.057Z")

@Validated
@Api(value = "transaction", description = "the transaction API")
@RequestMapping(value = "")
public interface TransactionApi {

	@CrossOrigin("*")
	@ApiOperation(value = "make a transaction between accounts", nickname = "makeTransaction", notes = "", response = ModelApiResponse.class, tags = {
			"transactions", })
	@ApiResponses(value = {
			@ApiResponse(code = 200, message = "Success creating a bank account.", response = ModelApiResponse.class),
			@ApiResponse(code = 405, message = "Invalid input") })
	@RequestMapping(value = "/transaction", produces = { "application/json" }, consumes = {
			"application/json" }, method = RequestMethod.POST)
	ResponseEntity<ModelApiResponse> makeTransaction(
			@ApiParam(value = "", required = true) @Valid @RequestBody Transaction body);

	@CrossOrigin("*")
	@ApiOperation(value = "Find transactions between two dates", nickname = "getTransactions", notes = "", response = Transaction.class, responseContainer = "List", tags = {
			"transactions", })
	@ApiResponses(value = {
			@ApiResponse(code = 200, message = "successful operation", response = Transaction.class, responseContainer = "List"),
			@ApiResponse(code = 400, message = "Invalid date supplied"),
			@ApiResponse(code = 404, message = "No Transaction found") })
	@RequestMapping(value = "/transaction", produces = { "application/json" }, method = RequestMethod.GET)
	ResponseEntity<List<Transaction>> getTransactions(
			@ApiParam(value = "the start date from where to start the search") @Valid @RequestParam(value = "start", required = false) String start,
			@ApiParam(value = "the end date of the search") @Valid @RequestParam(value = "end", required = false) String end) throws ParseException;

}